Classify Integration Tests
- Create separated packages to classify the integration tests depending on the functionality being tested
- Organize tests in the new packages
- Rename some classes in order to better represent what they're actually testing
