+++ submission/submission-webapp/src/itest/kotlin/ac/uk/ebi/biostd/itest/security/SecurityTest.kt
@@ -1,4 +1,4 @@
-package ac.uk.ebi.biostd.itest
+package ac.uk.ebi.biostd.itest.security
 
 import ac.uk.ebi.biostd.client.integration.web.SecurityWebClient
 import ac.uk.ebi.biostd.common.config.PersistenceConfig
